fix: token summary by codetype

+++ model/do.go
@@ -57,7 +57,7 @@ type BlockTokenVolumeDO struct {
 	Height       uint32 `db:"height"` // 区块高度
 	CodeHash     []byte `db:"codehash"`
 	Genesis      []byte `db:"genesis"`
-	CodeType     uint8  `db:"code_type"`      // 合约类型 0: nft, 1: ft
+	CodeType     uint32 `db:"code_type"`      // 合约类型 0: None, 1: FT, 2: UNIQUE, 3: NFT
 	NFTIdx       uint64 `db:"nft_idx"`        // nft tokenIdx
 	InDataValue  uint64 `db:"in_data_value"`  // 输入数量
 	OutDataValue uint64 `db:"out_data_value"` // 输出数量
